Consolidate OTEL into mc-telemetry
