+++ Cargo.toml
@@ -86,7 +86,6 @@ mp-chain-config = { path = "madara/crates/primitives/chain_config", default-feat
 mp-oracle = { path = "madara/crates/primitives/oracle", default-features = false }
 
 # Madara client
-mc-analytics = { path = "madara/crates/client/analytics" }
 mc-class-exec = { path = "madara/crates/client/cairo_native" }
 mc-telemetry = { path = "madara/crates/client/telemetry" }
 mc-db = { path = "madara/crates/client/db" }
@@ -155,7 +154,6 @@ opentelemetry-otlp = { version = "0.30.0", features = [
 ] }
 opentelemetry-semantic-conventions = { version = "0.30.0" }
 opentelemetry-appender-tracing = { version = "0.30.0", default-features = false }
-prometheus = "0.13.4"
 tracing = { version = "0.1.40", default-features = false }
 tracing-core = { version = "0.1.32", default-features = false }
 tracing-subscriber = { version = "0.3.18", features = [
