--- madara/crates/client/analytics/Cargo.toml
@@ -1,45 +1,0 @@
-[package]
-name = "mc-analytics"
-description = "Madara client open telemetry analytics service"
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-repository.workspace = true
-version.workspace = true
-homepage.workspace = true
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[dependencies]
-mp-utils = { workspace = true }
-
-anyhow = { workspace = true }
-async-trait = { workspace = true }
-console = "0.15" # for styled output
-futures = { workspace = true }
-http-body-util.workspace = true
-hyper = { workspace = true, features = ["server"] }
-hyper-util.workspace = true
-opentelemetry = { workspace = true, features = ["metrics", "logs"] }
-opentelemetry-appender-tracing = { workspace = true, default-features = false }
-opentelemetry-otlp = { workspace = true, features = [
-  "grpc-tonic",
-  "metrics",
-  "logs",
-] }
-opentelemetry-semantic-conventions = { workspace = true }
-opentelemetry-stdout = { workspace = true }
-opentelemetry_sdk = { workspace = true, features = ["rt-tokio", "logs"] }
-prometheus = { workspace = true }
-serde_json = { workspace = true }
-time = { workspace = true, features = ["formatting", "local-offset", "macros"] }
-tokio = { workspace = true }
-tracing = { workspace = true }
-tracing-core = { workspace = true, default-features = false }
-tracing-opentelemetry = { workspace = true }
-tracing-subscriber = { workspace = true, features = ["env-filter"] }
-url = { workspace = true }
-
-[dev-dependencies]
-mp-utils = { workspace = true }
