+++ madara/crates/client/block_production/src/metrics.rs
@@ -1,5 +1,5 @@
 use crate::util::ExecutionStats;
-use mc_analytics::{
+use mc_telemetry::{
     register_counter_metric_instrument, register_gauge_metric_instrument, register_histogram_metric_instrument,
 };
 use opentelemetry::metrics::{Counter, Gauge, Histogram};
