+++ madara/crates/client/cairo_native/Cargo.toml
@@ -17,7 +17,7 @@ targets = ["x86_64-unknown-linux-gnu"]
 [dependencies]
 
 # Madara
-mc-analytics = { workspace = true }
+mc-telemetry = { workspace = true }
 mp-class = { workspace = true }
 mp-convert = { workspace = true }
