+++ madara/crates/client/db/Cargo.toml
@@ -19,7 +19,7 @@ targets = ["x86_64-unknown-linux-gnu"]
 [dependencies]
 
 # Madara
-mc-analytics = { workspace = true }
+mc-telemetry = { workspace = true }
 mc-class-exec = { workspace = true }
 mp-block = { workspace = true }
 mp-bloom-filter = { workspace = true }
