+++ madara/crates/client/exec/Cargo.toml
@@ -17,7 +17,7 @@ targets = ["x86_64-unknown-linux-gnu"]
 [dependencies]
 
 # Madara
-mc-analytics = { workspace = true }
+mc-telemetry = { workspace = true }
 mc-class-exec = { workspace = true }
 mc-db = { workspace = true }
 mp-block = { workspace = true }
