+++ madara/crates/client/sync/Cargo.toml
@@ -26,7 +26,7 @@ m-cairo-test-contracts.workspace = true
 [dependencies]
 
 # Madara
-mc-analytics.workspace = true
+mc-telemetry.workspace = true
 mc-db.workspace = true
 mc-gateway-client.workspace = true
 mc-settlement-client.workspace = true
