+++ madara/crates/client/telemetry/src/formatter.rs
@@ -14,7 +14,9 @@ use tracing_subscriber::{
 pub fn display_fn<F: Fn(&mut fmt::Formatter<'_>) -> fmt::Result>(f: F) -> impl fmt::Display {
     DisplayFromFn(f)
 }
+
 struct DisplayFromFn<F: Fn(&mut fmt::Formatter<'_>) -> fmt::Result>(F);
+
 impl<F: Fn(&mut fmt::Formatter<'_>) -> fmt::Result> fmt::Display for DisplayFromFn<F> {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         (self.0)(f)
@@ -35,6 +37,7 @@ struct RpcCallEventVisitor {
     res_len: Option<u64>,
     response_time: Option<u128>,
 }
+
 impl Visit for RpcCallEventVisitor {
     fn record_str(&mut self, field: &Field, value: &str) {
         if field.name() == "method" {
@@ -42,21 +45,25 @@ impl Visit for RpcCallEventVisitor {
             self.method.push_str(value);
         }
     }
+
     fn record_i64(&mut self, field: &Field, value: i64) {
         if field.name() == "status" {
             self.status = Some(value)
         }
     }
+
     fn record_u64(&mut self, field: &Field, value: u64) {
         if field.name() == "res_len" {
             self.res_len = Some(value)
         }
     }
+
     fn record_u128(&mut self, field: &Field, value: u128) {
         if field.name() == "response_time" {
             self.response_time = Some(value)
         }
     }
+
     fn record_debug(&mut self, _field: &Field, _value: &dyn fmt::Debug) {
         // ignored
     }
@@ -69,6 +76,7 @@ impl RpcCallEventVisitor {
         self.res_len = None;
         self.response_time = None;
     }
+
     pub fn get(&self) -> Option<RpcCallEvent<'_>> {
         if self.method.is_empty() {
             return None;
@@ -99,7 +107,6 @@ impl Visit for CairoNativeEventVisitor {
     fn record_debug(&mut self, field: &Field, value: &dyn fmt::Debug) {
         match field.name() {
             "message" => {
-                // Remove quotes from Debug formatting
                 let formatted = format!("{:?}", value);
                 self.message = formatted.trim_matches('"').to_string();
             }
@@ -108,7 +115,6 @@ impl Visit for CairoNativeEventVisitor {
                 self.class_hash = Some(formatted.trim_matches('"').to_string());
             }
             "elapsed" | "duration" | "compile" | "load" | "convert" | "total" => {
-                // Format Duration nicely (remove quotes, keep the formatted duration)
                 let formatted = format!("{:?}", value);
                 self.elapsed = Some(formatted.trim_matches('"').to_string());
             }
@@ -160,7 +166,6 @@ impl Visit for CairoNativeEventVisitor {
     }
 }
 
-/// Visitor that collects all fields from a close_block event into a JSON object
 #[derive(Default)]
 struct CloseBlockEventVisitor {
     fields: Vec<(String, serde_json::Value)>,
@@ -177,7 +182,6 @@ impl Visit for CloseBlockEventVisitor {
     }
 
     fn record_u128(&mut self, field: &Field, value: u128) {
-        // JSON doesn't support u128 natively, convert to string if too large
         if value <= u64::MAX as u128 {
             self.fields.push((field.name().to_string(), serde_json::Value::Number((value as u64).into())));
         } else {
@@ -255,6 +259,7 @@ impl CairoNativeEventVisitor {
 
 pub fn visit_message(event: &tracing::Event<'_>, f: impl FnOnce(&dyn fmt::Debug) -> fmt::Result) -> fmt::Result {
     struct Visitor<F>(Option<F>, fmt::Result);
+
     impl<F: FnOnce(&dyn fmt::Debug) -> fmt::Result> Visit for Visitor<F> {
         fn record_debug(&mut self, field: &Field, value: &dyn fmt::Debug) {
             if field.name() == "message" {
@@ -264,6 +269,7 @@ pub fn visit_message(event: &tracing::Event<'_>, f: impl FnOnce(&dyn fmt::Debug)
             }
         }
     }
+
     let mut visitor = Visitor(Some(f), Ok(()));
     event.record(&mut visitor);
     visitor.1
@@ -294,7 +300,6 @@ impl CustomFormatter {
         display_fn(|f| {
             let datetime: OffsetDateTime = (*ts).into();
             let local_datetime = datetime.to_offset(self.local_offset);
-            // This allocates a String also :/
             match local_datetime.format(&self.ts_format) {
                 Ok(ts) => {
                     write!(f, "{}{}{}", self.open_bracket_dim, self.dim_style.apply_to(ts), self.closed_bracket_dim)
@@ -360,16 +365,12 @@ impl CustomFormatter {
         let error = visitor.get_error();
         let native_enabled = visitor.get_native_enabled();
 
-        // Darker cyan for CAIRO_NATIVE prefix (more subtle)
         let cairo_native_prefix = Style::new().cyan().dim().apply_to("CAIRO_NATIVE");
-        // Muted color for class_hash (less prominent than CAIRO_NATIVE)
         let class_hash_style = Style::new().dim();
         let error_style = Style::new().red();
 
-        // Format: timestamp + CAIRO_NATIVE + [WARN/ERROR] + message + class_hash + native_enabled + (optional error) + (optional timing)
         write!(writer, "{} {}", self.timestamp_fmt(ts), cairo_native_prefix)?;
 
-        // Level prefix shown for WARN and ERROR only
         match *level {
             Level::WARN => {
                 write!(writer, " {}", Style::new().yellow().apply_to("WARN"))?;
@@ -377,21 +378,17 @@ impl CustomFormatter {
             Level::ERROR => {
                 write!(writer, " {}", Style::new().red().apply_to("ERROR"))?;
             }
-            _ => {} // INFO and DEBUG don't show level prefix
+            _ => {}
         }
 
-        // Message formatting improves clarity by capitalizing and replacing underscores
         let formatted_message = Self::format_message(message);
         write!(writer, " {}", formatted_message)?;
 
         if let Some(hash) = class_hash {
-            // class_hash displayed with full hash value, less prominent styling
-            // Quotes removed if present from Debug formatting
             let hash_clean = hash.trim_matches('"');
             write!(writer, " {}", class_hash_style.apply_to(format!("class_hash={}", hash_clean)))?;
         }
 
-        // Display native_enabled if present
         if let Some(enabled) = native_enabled {
             write!(writer, " {}", class_hash_style.apply_to(format!("native_enabled={}", enabled)))?;
         }
@@ -400,7 +397,6 @@ impl CustomFormatter {
             write!(writer, " {}", error_style.apply_to(format!("error={}", err)))?;
         }
 
-        // Format timing - prefer elapsed_ms if available, otherwise use elapsed Duration string
         let timing_str = Self::format_timing(elapsed_ms, conversion_ms, load_ms, convert_ms, elapsed);
         if let Some(timing) = timing_str {
             let elapsed_style = Self::get_timing_style(&timing);
@@ -410,8 +406,6 @@ impl CustomFormatter {
         writeln!(writer)
     }
 
-    /// Format timing information from various sources
-    /// Prefers elapsed_ms if available, shows breakdown if multiple components exist
     fn format_timing(
         elapsed_ms: Option<u64>,
         conversion_ms: Option<u64>,
@@ -419,16 +413,12 @@ impl CustomFormatter {
         convert_ms: Option<u64>,
         elapsed: Option<&str>,
     ) -> Option<String> {
-        // If we have elapsed_ms, use it as primary timing
         if let Some(total_ms) = elapsed_ms {
             let mut parts = Vec::new();
-
-            // Format total time
             let total_str =
                 if total_ms >= 1000 { format!("{:.3}s", total_ms as f64 / 1000.0) } else { format!("{}ms", total_ms) };
             parts.push(total_str);
 
-            // Add breakdown if we have component timings
             let mut breakdown = Vec::new();
             if let Some(load) = load_ms {
                 breakdown.push(format!("load: {}ms", load));
@@ -447,19 +437,13 @@ impl CustomFormatter {
             return Some(parts.join(" "));
         }
 
-        // Fall back to elapsed Duration string if available
         elapsed.map(|s| s.to_string())
     }
 
-    /// Format message for better clarity
     fn format_message(message: &str) -> String {
-        // Remove quotes if present from Debug formatting
         let cleaned = message.trim_matches('"');
+        let spaced = cleaned.replace('_', " ");
 
-        // Replace underscores with spaces for readability
-        let spaced = cleaned.replace("_", " ");
-
-        // Capitalize first letter
         if let Some(first) = spaced.chars().next() {
             format!("{}{}", first.to_uppercase(), &spaced[1..])
         } else {
@@ -467,42 +451,33 @@ impl CustomFormatter {
         }
     }
 
-    /// Get style for timing based on duration value
-    /// Duration Debug format: "15.833µs", "1.732s", "234ms", "123ns"
     fn get_timing_style(timing_str: &str) -> Style {
-        // Parse timing string to extract numeric value and unit
         let timing_clean = timing_str.trim();
 
-        // Handle different time units (check longer units first)
         if timing_clean.ends_with("ns") {
-            // Nanoseconds - always very fast, use dim
             return Style::new().dim();
         } else if timing_clean.ends_with("µs") || timing_clean.ends_with("us") {
-            // Microseconds - always fast, use dim
             return Style::new().dim();
         } else if timing_clean.ends_with("ms") {
-            // Milliseconds
             if let Ok(val) = timing_clean.trim_end_matches("ms").trim().parse::<f64>() {
                 if val > 1000.0 {
-                    return Style::new().red(); // Very slow (>1s)
+                    return Style::new().red();
                 } else if val > 100.0 {
-                    return Style::new().yellow(); // Slow (>100ms)
+                    return Style::new().yellow();
                 }
             }
-            return Style::new().dim(); // <100ms, normal
-        } else if timing_clean.ends_with("s") {
-            // Seconds - yellow/red for slower operations
-            if let Ok(val) = timing_clean.trim_end_matches("s").trim().parse::<f64>() {
+            return Style::new().dim();
+        } else if timing_clean.ends_with('s') {
+            if let Ok(val) = timing_clean.trim_end_matches('s').trim().parse::<f64>() {
                 if val > 5.0 {
-                    return Style::new().red(); // Very slow (>5s)
+                    return Style::new().red();
                 } else if val > 1.0 {
-                    return Style::new().yellow(); // Slow (>1s)
+                    return Style::new().yellow();
                 }
             }
-            return Style::new().dim(); // <1s, normal
+            return Style::new().dim();
         }
 
-        // Default: dim for unknown format
         Style::new().dim()
     }
 
@@ -517,7 +492,6 @@ impl CustomFormatter {
         let mut visitor = RpcCallEventVisitor::default();
         event.record(&mut visitor);
         let Some(rpc_call_event) = visitor.get() else {
-            // Fallback to normal formatter.
             return self.format_with_target(writer, event, target, ts, level, &Style::new().blue());
         };
 
@@ -537,7 +511,6 @@ impl CustomFormatter {
                 rpc_call_event.method,
                 status_style.apply_to(&rpc_call_event.status),
                 rpc_call_event.res_len,
-                // Conversion from micros u128 to u64 should be fine.
                 time_style.apply_to(&Duration::from_micros(rpc_call_event.response_time as u64)),
             )
         } else {
@@ -549,13 +522,11 @@ impl CustomFormatter {
                 rpc_call_event.method,
                 status_style.apply_to(&rpc_call_event.status),
                 rpc_call_event.res_len,
-                // Convertion from micros u128 to u64 should be fine.
                 time_style.apply_to(&Duration::from_micros(rpc_call_event.response_time as u64)),
             )
         }
     }
 
-    /// Format close_block events as JSON for Loki ingestion
     fn format_close_block(
         &self,
         writer: &mut Writer<'_>,
@@ -567,7 +538,6 @@ impl CustomFormatter {
         let mut visitor = CloseBlockEventVisitor::default();
         event.record(&mut visitor);
 
-        // Build JSON object with timestamp, level, target, and all fields
         let datetime: OffsetDateTime = (*ts).into();
         let local_datetime = datetime.to_offset(self.local_offset);
         let timestamp_str = local_datetime
@@ -580,13 +550,11 @@ impl CustomFormatter {
         json_obj.insert("target".to_string(), serde_json::Value::String(target.to_string()));
         json_obj.insert("message".to_string(), serde_json::Value::String(visitor.message.clone()));
 
-        // Add all captured fields
         for (key, value) in visitor.fields {
             json_obj.insert(key, value);
         }
 
         let json_str = serde_json::to_string(&serde_json::Value::Object(json_obj)).unwrap_or_else(|_| "{}".to_string());
-
         writeln!(writer, "{}", json_str)
     }
 }
@@ -603,7 +571,6 @@ where
         event: &tracing::Event<'_>,
     ) -> std::fmt::Result {
         let ts = SystemTime::now();
-
         let metadata = event.metadata();
         let level = metadata.level();
         let target = metadata.target();
