+++ madara/crates/primitives/block/Cargo.toml
@@ -21,7 +21,7 @@ num-traits = { workspace = true }
 # Madara
 bitvec = { workspace = true }
 bonsai-trie = { workspace = true }
-mc-analytics = { workspace = true }
+mc-telemetry = { workspace = true }
 mp-chain-config = { workspace = true }
 mp-convert = { workspace = true }
 mp-receipt = { workspace = true }
