+++ madara/crates/primitives/utils/src/service.rs
@@ -377,7 +377,6 @@ pub enum MadaraServiceId {
     RpcAdmin,
     Gateway,
     Telemetry,
-    Analytics,
     Mempool,
     ExternalDb,
 }
@@ -395,9 +394,8 @@ impl ServiceId for MadaraServiceId {
             MadaraServiceId::RpcAdmin => PowerOfTwo::P5,
             MadaraServiceId::Gateway => PowerOfTwo::P6,
             MadaraServiceId::Telemetry => PowerOfTwo::P7,
-            MadaraServiceId::Analytics => PowerOfTwo::P8,
-            MadaraServiceId::Mempool => PowerOfTwo::P9,
-            MadaraServiceId::ExternalDb => PowerOfTwo::P10,
+            MadaraServiceId::Mempool => PowerOfTwo::P8,
+            MadaraServiceId::ExternalDb => PowerOfTwo::P9,
         }
     }
 }
@@ -417,7 +415,6 @@ impl Display for MadaraServiceId {
                 Self::RpcAdmin => "rpc admin",
                 Self::Gateway => "gateway",
                 Self::Telemetry => "telemetry",
-                Self::Analytics => "analytics",
                 Self::Mempool => "mempool",
                 Self::ExternalDb => "external db",
             }
@@ -453,9 +450,8 @@ impl From<PowerOfTwo> for MadaraServiceId {
             PowerOfTwo::P5 => Self::RpcAdmin,
             PowerOfTwo::P6 => Self::Gateway,
             PowerOfTwo::P7 => Self::Telemetry,
-            PowerOfTwo::P8 => Self::Analytics,
-            PowerOfTwo::P9 => Self::Mempool,
-            PowerOfTwo::P10 => Self::ExternalDb,
+            PowerOfTwo::P8 => Self::Mempool,
+            PowerOfTwo::P9 => Self::ExternalDb,
             _ => Self::Monitor, // Default fallback for unknown values
         }
     }
@@ -599,7 +595,7 @@ impl MadaraServiceMask {
     }
 
     fn active_set(&self) -> Vec<MadaraServiceId> {
-        let mut i = MadaraServiceId::Telemetry.svc_id() as u64;
+        let mut i = MadaraServiceId::ExternalDb.svc_id() as u64;
         let state = self.value();
         let mut set = Vec::with_capacity(SERVICE_COUNT_MAX);
 
@@ -1031,10 +1027,10 @@ pub struct ServiceTransport {
 ///     #[inline(always)]
 ///     fn svc_id(&self) -> PowerOfTwo {
 ///         match self {
-///             // PowerOfTwo::P0 up until PowerOfTwo::P7 are already in use by
+///             // PowerOfTwo::P0 up until PowerOfTwo::P9 are already in use by
 ///             // MadaraServiceId, you should not use them!
-///             Self::MyServiceA => PowerOfTwo::P8,
-///             Self::MyServiceB => PowerOfTwo::P9,
+///             Self::MyServiceA => PowerOfTwo::P10,
+///             Self::MyServiceB => PowerOfTwo::P11,
 ///         }
 ///     }
 /// }
