+++ madara/crates/tests/src/lib.rs
@@ -563,8 +563,6 @@ impl MadaraCmdBuilder {
                     .into_iter()
                     .flatten(),
             )
-            // Always use dynamic ports for services that may conflict when running multiple nodes
-            .args(["--analytics-prometheus-endpoint-port", "0"])
             .args(["--rpc-admin-port", "0"])
             .args(gateway_key_args)
             .stdout(Stdio::piped())
@@ -592,6 +590,37 @@ fn madara_help_shows() {
     assert!(output.status.success());
     let stdout = String::from_utf8(output.stdout).unwrap();
     assert!(stdout.contains("Madara: High performance Starknet sequencer/full-node"), "stdout: {stdout}");
+}
+
+#[rstest]
+#[tokio::test]
+async fn madara_starts_without_otel_endpoint() {
+    let _ = tracing_subscriber::fmt().with_test_writer().try_init();
+
+    let mut node = MadaraCmdBuilder::new()
+        .args(["--devnet", "--no-l1-sync", "--l1-gas-price", "0", "--blob-gas-price", "0"])
+        .run();
+    node.wait_for_ready().await;
+}
+
+#[rstest]
+#[tokio::test]
+async fn madara_starts_with_otel_endpoint() {
+    let _ = tracing_subscriber::fmt().with_test_writer().try_init();
+
+    let mut node = MadaraCmdBuilder::new()
+        .args([
+            "--devnet",
+            "--no-l1-sync",
+            "--l1-gas-price",
+            "0",
+            "--blob-gas-price",
+            "0",
+            "--otel-collector-endpoint",
+            "http://127.0.0.1:4317",
+        ])
+        .run();
+    node.wait_for_ready().await;
 }
 
 #[rstest]
