+++ madara/node/src/cli/telemetry.rs
@@ -1,41 +1,98 @@
-use clap::Args;
+use clap::{ArgAction, Args, Parser};
+use mc_telemetry::TelemetryConfig;
+use mp_utils::parsers::parse_url;
 use serde::{Deserialize, Serialize};
+use url::Url;
 
-/// Parameters used to config telemetry.
+fn default_service_name() -> String {
+    "madara".into()
+}
+
+fn default_metrics_export() -> bool {
+    true
+}
+
+/// Parameters used to configure OpenTelemetry.
 #[derive(Debug, Clone, Args, Deserialize, Serialize)]
 pub struct TelemetryParams {
-    /// Enable connecting to the Madara telemetry server.
-    #[arg(env = "MADARA_TELEMETRY", long)]
-    pub telemetry: bool,
+    /// Name of the OTEL service.
+    #[arg(env = "MADARA_OTEL_SERVICE_NAME", long, default_value = default_service_name())]
+    #[serde(default = "default_service_name")]
+    pub otel_service_name: String,
 
-    /// The URL of the telemetry server.
-    /// Pass this flag multiple times specify multiple telemetry endpoints.
-    /// Verbosity levels range from 0-9, with 0 denoting
-    /// the least verbosity.
-    /// Expected format is 'URL VERBOSITY', e.g. `--telemetry-url 'wss://foo/bar 0'`.
-    #[arg(
-		env = "MADARA_TELEMETRY_URL",
-		long = "telemetry-url",
-		value_name = "URL VERBOSITY",
-		value_parser = parse_telemetry_endpoints,
-		default_value = "wss://starknodes.com/submit 0",
-	)]
-    pub telemetry_endpoints: Vec<(String, u8)>,
+    /// Endpoint of the OTEL collector.
+    #[arg(env = "OTEL_EXPORTER_OTLP_ENDPOINT", long, value_parser = parse_url)]
+    #[serde(default)]
+    pub otel_collector_endpoint: Option<Url>,
+
+    /// Export metrics to the OTEL collector.
+    #[arg(env = "OTEL_EXPORT_METRICS", long, action = ArgAction::Set, default_value_t = default_metrics_export())]
+    #[serde(default = "default_metrics_export")]
+    pub otel_export_metrics: bool,
+
+    /// Export traces to the OTEL collector.
+    #[arg(env = "OTEL_EXPORT_TRACES", long)]
+    #[serde(default)]
+    pub otel_export_traces: bool,
+
+    /// Export logs to the OTEL collector.
+    #[arg(env = "OTEL_EXPORT_LOGS", long)]
+    #[serde(default)]
+    pub otel_export_logs: bool,
 }
 
-#[derive(Debug, thiserror::Error)]
-enum TelemetryParsingError {
-    #[error("verbosity must be an int")]
-    VerbosityParsingError(std::num::ParseIntError),
-}
-
-fn parse_telemetry_endpoints(s: &str) -> Result<(String, u8), TelemetryParsingError> {
-    match s.find(' ') {
-        None => Ok((s.to_string(), 1)),
-        Some(pos) => {
-            let url = s[..pos].to_string();
-            let verbosity = s[pos + 1..].parse().map_err(TelemetryParsingError::VerbosityParsingError)?;
-            Ok((url, verbosity))
+impl TelemetryParams {
+    pub fn as_telemetry_config(&self) -> TelemetryConfig {
+        TelemetryConfig {
+            service_name: self.otel_service_name.clone(),
+            collection_endpoint: self.otel_collector_endpoint.clone(),
+            export_metrics: self.otel_export_metrics,
+            export_traces: self.otel_export_traces,
+            export_logs: self.otel_export_logs,
         }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::cli::RunCmd;
+
+    #[test]
+    fn telemetry_params_use_new_defaults() {
+        let run_cmd = RunCmd::parse_from(["madara", "--full", "--network", "sepolia"]);
+
+        assert_eq!(run_cmd.telemetry_params.otel_service_name, "madara");
+        assert_eq!(run_cmd.telemetry_params.otel_collector_endpoint, None);
+        assert!(run_cmd.telemetry_params.otel_export_metrics);
+        assert!(!run_cmd.telemetry_params.otel_export_traces);
+        assert!(!run_cmd.telemetry_params.otel_export_logs);
+    }
+
+    #[test]
+    fn telemetry_params_parse_explicit_otel_flags() {
+        let run_cmd = RunCmd::parse_from([
+            "madara",
+            "--full",
+            "--network",
+            "sepolia",
+            "--otel-service-name",
+            "custom-node",
+            "--otel-collector-endpoint",
+            "http://127.0.0.1:4317",
+            "--otel-export-metrics",
+            "false",
+            "--otel-export-traces",
+            "--otel-export-logs",
+        ]);
+
+        assert_eq!(run_cmd.telemetry_params.otel_service_name, "custom-node");
+        assert_eq!(
+            run_cmd.telemetry_params.otel_collector_endpoint.as_ref().map(Url::as_str),
+            Some("http://127.0.0.1:4317/")
+        );
+        assert!(!run_cmd.telemetry_params.otel_export_metrics);
+        assert!(run_cmd.telemetry_params.otel_export_traces);
+        assert!(run_cmd.telemetry_params.otel_export_logs);
     }
 }
