+++ madara/node/src/main.rs
@@ -20,7 +20,6 @@
 //! These are crates which are responsible for the core (business) functionality of the node. They
 //! answer the question of _how_ the node transforms data.
 //!
-//! - [mc-analytics]
 //! - [mc-block-production]
 //! - [mc-db]
 //! - [mc-devnet]
@@ -68,7 +67,6 @@
 //! [Starknet RPC specs]: https://github.com/starkware-libs/starknet-specs
 //! [Starknet P2P specs]: https://github.com/starknet-io/starknet-p2p-specs
 //!
-//! [mc-analytics]: mc_analytics
 //! [mc-block-production]: mc_block_production
 //! [mc-db]: mc_db
 //! [mc-devnet]: mc_devnet
@@ -116,7 +114,6 @@ use figment::{
     Figment,
 };
 use http::{HeaderName, HeaderValue};
-use mc_analytics::AnalyticsService;
 use mc_db::MadaraBackend;
 use mc_external_db::ExternalDbService;
 use mc_gateway_client::GatewayProvider;
@@ -175,10 +172,10 @@ async fn main() -> anyhow::Result<()> {
     let mut run_cmd: RunCmd = config.extract()?;
     run_cmd.check_mode()?;
 
-    // Setting up analytics
-    let mut service_analytics = AnalyticsService::new(run_cmd.analytics_params.as_analytics_config())
-        .context("Initializing analytics service")?;
-    service_analytics.setup().context("Setting-up analystics service")?;
+    // Setting up telemetry
+    let mut service_telemetry = TelemetryService::new(run_cmd.telemetry_params.as_telemetry_config())
+        .context("Initializing telemetry service")?;
+    service_telemetry.setup().context("Setting-up telemetry service")?;
 
     // If it's a sequencer or a devnet we set the mandatory chain config. If it's a full node we set the chain config from the network or the custom chain config.
     let chain_config = if run_cmd.is_sequencer() {
@@ -209,8 +206,6 @@ async fn main() -> anyhow::Result<()> {
     }
 
     let node_name = run_cmd.node_name_or_provide().await.to_string();
-    let node_version = env!("MADARA_BUILD_VERSION");
-
     tracing::info!("🥷  {} Node", GREET_IMPL_NAME);
     tracing::info!("💁 Support URL: {}", GREET_SUPPORT_URL);
     tracing::info!("🏷  Node Name: {}", node_name);
@@ -245,12 +240,6 @@ async fn main() -> anyhow::Result<()> {
     // ===================================================================== //
     //                             SERVICES (SETUP)                          //
     // ===================================================================== //
-
-    // Telemetry
-
-    let service_telemetry: TelemetryService =
-        TelemetryService::new(run_cmd.telemetry_params.telemetry_endpoints.clone())
-            .context("Initializing telemetry service")?;
 
     // Database
 
@@ -309,10 +298,6 @@ async fn main() -> anyhow::Result<()> {
 
         if run_cmd.gateway_params.any_enabled() {
             deferred_service_start.push(MadaraServiceId::Gateway);
-        }
-
-        if run_cmd.telemetry_params.telemetry {
-            deferred_service_start.push(MadaraServiceId::Telemetry);
         }
 
         if run_cmd.is_sequencer() {
@@ -422,8 +407,6 @@ async fn main() -> anyhow::Result<()> {
     .await
     .context("Initializing gateway service")?;
 
-    service_telemetry.send_connected(&node_name, node_version, &chain_config.chain_name, &sys_info);
-
     // ===================================================================== //
     //                             SERVICES (START)                          //
     // ===================================================================== //
@@ -433,15 +416,14 @@ async fn main() -> anyhow::Result<()> {
     }
 
     let mut app = ServiceMonitor::default()
-        .with(service_analytics)?
+        .with(service_telemetry)?
         .with(service_mempool)?
         .with(service_l1_sync)?
         .with(service_l2_sync)?
         .with(service_block_production)?
         .with(service_rpc_user)?
         .with(service_rpc_admin)?
-        .with(service_gateway)?
-        .with(service_telemetry)?;
+        .with(service_gateway)?;
 
     if let Some(service_external_db) = service_external_db {
         app = app.with(service_external_db)?;
@@ -457,7 +439,7 @@ async fn main() -> anyhow::Result<()> {
     let warp_update_receiver = run_cmd.args_preset.warp_update_receiver;
 
     app.activate(MadaraServiceId::Mempool);
-    app.activate(MadaraServiceId::Analytics);
+    app.activate(MadaraServiceId::Telemetry);
 
     if l1_sync_enabled && (l1_endpoint_some || !run_cmd.devnet) {
         app.activate(MadaraServiceId::L1Sync);
@@ -481,10 +463,6 @@ async fn main() -> anyhow::Result<()> {
 
     if run_cmd.gateway_params.any_enabled() && !warp_update_receiver {
         app.activate(MadaraServiceId::Gateway);
-    }
-
-    if run_cmd.telemetry_params.telemetry && !warp_update_receiver {
-        app.activate(MadaraServiceId::Telemetry);
     }
 
     if external_db_configured && !warp_update_receiver {
